requirements.txt
python-telegram-bot==22.0
python-dotenv==1.1.0